backend/requirements.txt
asgiref==3.8.1
black==25.1.0
boto3==1.37.23
botocore==1.37.23
click==8.1.8
Django==4.2.20
django-cors-headers==4.7.0
django-filter==24.1
django-storages==1.14.5
djangorestframework==3.15.2
exceptiongroup==1.2.2
gunicorn==23.0.0
iniconfig==2.0.0
isort==6.0.1
jmespath==1.0.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
Pillow==11.2.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg[binary]==3.2.6
pytest==8.3.5
python-dateutil==2.9.0.post0
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.1
urllib3==1.26.18
python-decouple